shared: change serialization to use size_t
Had a few missing places where size_t wasn't being used to store
serialized data (and then had conversions when deserializing) that
could cause problems so switch to using size_t.

+++ src/shared/serialize.h
@@ -85,7 +85,7 @@ bool buxton_deserialize(uint8_t *source, BuxtonData *dest);
  * @return a size_t, 0 indicates failure otherwise size of dest
  */
 size_t buxton_serialize_message(uint8_t **dest, BuxtonControlMessage message,
-			      unsigned int n_params, ...);
+			      size_t n_params, ...);
 
 /**
  * Deserialize the given data into an array of BuxtonData structs
